Separated the Connection Menu code from the connection itself
